refactor: adding cascade deletion of product to shopping cart items, changing integer types to BIGINT

+++ server/db/models/user.js
@@ -26,7 +26,7 @@ const User = db.define('user', {
     type: Sequelize.STRING
   },
   phoneNumber: {
-    type: Sequelize.INTEGER // PHONE NUMBERRRRR
+    type: Sequelize.BIGINT
   },
   firstName: {
     type: Sequelize.STRING
